src/tax_assistant/requirements.txt
django
lxml
requests
django_rest_framework
openai
uwsgi
django-cors-headers
django-extensions
psycopg2-binary
